--- test/Interop/Cxx/class/Inputs/synthesized-initializers.h
@@ -1,5 +1,0 @@
-struct EmptyStruct {};
-
-struct IntBox {
-  int x;
-};
